Fixes for queue storage (#11)
* Fixes for queue storage:

1) fix for how the canonicalized headers are constructed: based on the documentation, each HTTP header name should be converted to lowercase and, any white space around the colon in the header should be trimmed.

2) fix the retrieve of the queue metadata
When trying to get the metadata for a queue with the following metadata:

{
    clientName: 'fast-azure-storage'
}

the response headers will look something like this:

Transfer-Encoding: chunked
x-ms-clientname: fast-azure-storage
Date: Fri, 16 Sep 2011 01:27:38 GMT
Server: Windows-Azure-Queue/1.0 Microsoft-HTTPAPI/2.0

Meaning that the result of the getMetadata will be:
{
    clientname: 'fast-azure-storage'
}

This happens because the 'https' library returns the HTTP headers with lowercase. The solution was to read the header name from response.rawHeaders.

* - modifies the queue getMetadata test in order to demonstrate that the
  whitespace inside the header value is not an issue
- changes the way the metadata is read